Refactor pipeline_parallel code
Split the large pipeline_parallel.py file (3995 lines) into multiple
modules for better maintainability:

- pipeline_impl/__init__.py: Export all public interfaces
- pipeline_impl/utils.py: Utility functions and classes
- pipeline_impl/hook_system.py: Hook system
- pipeline_impl/p2p_handle.py: P2P communication
- pipeline_impl/no_pipeline.py: NoPipelineParallel class
- pipeline_impl/base_pipeline.py: PipelineParallel base class
- pipeline_impl/interleave_pipeline.py: PipelineParallelWithInterleave class
- pipeline_impl/fthenb_pipeline.py: PipelineParallelWithInterleaveFthenB class
- pipeline_impl/balanced_memory.py: VPPFhenBInBalancedMemory class

The original pipeline_parallel.py is now a simple redirect file that
re-exports all public interfaces from pipeline_impl sub-package,
ensuring backward compatibility.
